+++ cmd/neofs-node/config.go
@@ -25,6 +25,7 @@ import (
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/engine"
 	"github.com/nspcc-dev/neofs-node/pkg/metrics"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
+	balanceClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/balance"
 	cntClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/container"
 	nmClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/netmap"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
@@ -116,6 +117,7 @@ type basics struct {
 	cli  *client.Client
 	nCli *nmClient.Client
 	cCli *cntClient.Client
+	bCli *balanceClient.Client
 
 	// all caches are non-nil iff caching is enabled in config
 	containerCache     *ttlContainerStorage
@@ -198,6 +200,7 @@ type cfg struct {
 	cfgGRPC           cfgGRPC
 	cfgMeta           cfgMeta
 	cfgMorph          cfgMorph
+	cfgBalance        cfgBalance
 	cfgContainer      cfgContainer
 	cfgNodeInfo       cfgNodeInfo
 	cfgNetmap         cfgNetmap
@@ -241,7 +244,13 @@ type cfgMorph struct {
 	proxyScriptHash neogoutil.Uint160
 }
 
+type cfgBalance struct {
+	parsers     map[event.Type]event.NotificationParser
+	subscribers map[event.Type][]event.Handler
+}
+
 type cfgContainer struct {
+	payments    *paymentChecker
 	parsers     map[event.Type]event.NotificationParser
 	subscribers map[event.Type][]event.Handler
 	workerPool  util.WorkerPool // pool for asynchronous handlers
@@ -410,6 +419,10 @@ func initCfg(appCfg *config.Config) *cfg {
 		persistate:        persistate,
 		privateTokenStore: persistate,
 	}
+	c.cfgBalance = cfgBalance{
+		parsers:     make(map[event.Type]event.NotificationParser),
+		subscribers: make(map[event.Type][]event.Handler),
+	}
 	c.cfgContainer = cfgContainer{
 		workerPool: containerWorkerPool,
 	}
@@ -509,6 +522,9 @@ func initBasics(c *cfg, key *keys.PrivateKey, stateStorage *state.PersistentStor
 	currBlock, err := cli.BlockCount()
 	fatalOnErr(err)
 	nState.block.Store(currBlock)
+
+	b.bCli, err = balanceClient.NewFromMorph(cli, b.balanceSH)
+	fatalOnErr(err)
 
 	b.cCli, err = cntClient.NewFromMorph(cli, b.containerSH)
 	fatalOnErr(err)