+++ cmd/neofs-node/container.go
@@ -9,8 +9,10 @@ import (
 	"time"
 
 	containerCore "github.com/nspcc-dev/neofs-node/pkg/core/container"
+	balanceClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/balance"
 	cntClient "github.com/nspcc-dev/neofs-node/pkg/morph/client/container"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
+	balanceEvent "github.com/nspcc-dev/neofs-node/pkg/morph/event/balance"
 	containerEvent "github.com/nspcc-dev/neofs-node/pkg/morph/event/container"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event/netmap"
 	containerService "github.com/nspcc-dev/neofs-node/pkg/services/container"
@@ -85,6 +87,7 @@ func initContainerService(c *cfg) {
 		})
 	}
 
+	initPaymentChecker(c)
 	initSizeLoadReports(c)
 
 	cnrSrv := containerService.New(&c.key.PrivateKey, c.networkState, c.cli, (*containersInChain)(&c.basics), c.nCli)
@@ -349,6 +352,82 @@ func setContainerNotificationParser(c *cfg, sTyp string, p event.NotificationPar
 	}
 
 	c.cfgContainer.parsers[typ] = p
+}
+
+func initPaymentChecker(c *cfg) {
+	var (
+		l = c.log.With(zap.String("component", "paymentChecker"))
+		p = &paymentChecker{
+			m:          sync.RWMutex{},
+			statuses:   make(map[cid.ID]int64),
+			balanceCli: c.bCli,
+		}
+	)
+
+	const changeUnpaidStatusEventName = "ChangePaymentStatus"
+	typ := event.TypeFromString(changeUnpaidStatusEventName)
+	c.cfgBalance.parsers[typ] = balanceEvent.ParseChangePaymentStatus
+	c.cfgBalance.subscribers[typ] = append(c.cfgBalance.subscribers[typ], func(e event.Event) {
+		ev := e.(balanceEvent.ChangePaymentStatus)
+		p.m.Lock()
+		defer p.m.Unlock()
+
+		cID := cid.ID(ev.ContainerID)
+		if ev.Unpaid {
+			l.Info("container status has changed to unpaid",
+				zap.Stringer("cID", cID),
+				zap.Uint64("epoch", ev.Epoch))
+
+			p.statuses[cID] = int64(ev.Epoch)
+		} else {
+			l.Info("container status has changed to paid",
+				zap.Stringer("cID", cID),
+				zap.Uint64("epoch", ev.Epoch))
+
+			p.statuses[cID] = -1
+		}
+	})
+
+	c.cfgContainer.payments = p
+}
+
+type paymentChecker struct {
+	m        sync.RWMutex
+	statuses map[cid.ID]int64
+
+	balanceCli *balanceClient.Client
+}
+
+func (p *paymentChecker) resetCache() {
+	p.m.Lock()
+	defer p.m.Unlock()
+
+	clear(p.statuses)
+}
+
+func (p *paymentChecker) UnpaidSince(cID cid.ID) (int64, error) {
+	p.m.RLock()
+	epoch, ok := p.statuses[cID]
+	p.m.RUnlock()
+
+	if ok {
+		return epoch, nil
+	}
+
+	p.m.Lock()
+	defer p.m.Unlock()
+	epoch, ok = p.statuses[cID]
+	if ok {
+		return epoch, nil
+	}
+
+	epoch, err := p.balanceCli.GetUnpaidContainerEpoch(cID)
+	if err != nil {
+		return 0, fmt.Errorf("FS chain RPC call: %w", err)
+	}
+	p.statuses[cID] = epoch
+
+	return epoch, nil
 }
 
 func (c *cfg) PublicKey() []byte {