+++ cmd/neofs-node/main.go
@@ -209,6 +209,7 @@ func (c *cfg) restartMorph() error {
 	c.log.Info("restarting internal services because of RPC connection loss...")
 
 	c.resetCaches()
+	c.cfgContainer.payments.resetCache()
 
 	epoch, ni, err := getNetworkState(c)
 	if err != nil {