+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changelog for NeoFS Node
 - Move `persistent_sessions` data to `persistent_state` in SN (#3630)
 - More efficient write cache batching for small objects (#3414)
 - SN now indexes metadata for root objects nested twice (parents of parents) (#3674)
+- Storage nodes do not serve unpaid containers (#774)
 
 ### Removed
 - Pre-0.46.0 write cache format migration (#3647)